usageQuota middleware writing to couch
